fix(vrf): rename package + review comments
